purchase_order_forecast/requirements.txt
# Modules used
psycopg2-binary   # Module used for read and write from postgresql database via python
openpyxl         # Module used to write and read excel via python
pandas           # Module used to write and read excel via python